zp 2019-11-07-16:59

+++ app/src/main/java/com/ceshi/helloworld/bean/RequestSignBean.java
@@ -202,7 +202,7 @@ public class RequestSignBean {
             return pluAmount;
         }
 
-        public void setPluAmount(int pluAmount) {
+        public void setPluAmount(double pluAmount) {
             this.pluAmount = pluAmount;
         }
